requirements.txt
pytest>=7.0
requests>=2.28
PyYAML>=6.0
python-dotenv>=1.0
faker>=18.0
allure-pytest>=2.0.0
pytest-rerunfailures>=10.0
pytest-html>=3.2.0
requests-mock>=1.11.0
prometheus-client>=0.16.0    # optional: for metrics
pika>=1.3.0                  # RabbitMQ client (optional bonus)
